A2A/Embeddings/requirements.txt
streamlit==1.47.0
sentence-transformers==5.0.0
transformers==4.53.2
torch==2.7.1
faiss-cpu==1.11.0.post1
pandas==2.3.1
numpy==2.3.1

